refactor: 이동 전략 중복 코드 구현 분리
