+++ src/main/java/chess/domain/board/Position.java
@@ -33,6 +33,14 @@ public class Position {
                 .orElseThrow(RuntimeException::new);
     }
 
+    public int movedHorizontalWeight(int value) {
+        return this.horizontal.getWeight() + value;
+    }
+
+    public int movedVerticalWeight(int value) {
+        return this.vertical.getWeight() + value;
+    }
+
     public boolean isSameVertical(Vertical vertical) {
         return this.vertical == vertical;
     }
@@ -71,11 +79,6 @@ public class Position {
     public Position moveTowardDirection(Direction direction) {
         return Position.of(Horizontal.findFromWeight(getHorizontalWeight() + direction.getX()),
                 Vertical.findFromWeight(getVerticalWeight() + direction.getY()));
-    }
-
-    public boolean isDeadLine() {
-        Vertical vertical = getVertical();
-        return vertical == Vertical.ONE || vertical == Vertical.EIGHT;
     }
 
     public String convertToString() {
